+++ src/main/java/site/siredvin/progressiveperipherals/data/BlockTagsProvider.java
@@ -30,7 +30,8 @@ public class BlockTagsProvider extends TagsProvider<Block> {
                 Blocks.IRREALIUM_MACHINERY_STORAGE.get(),
                 Blocks.REALITY_BREAKTHROUGH_REACTOR_CONTROLLER.get(),
                 Blocks.REALITY_BREAKTHROUGH_EXTRACTOR_CONTROLLER.get(),
-                Blocks.IRREALIUM_MACHINERY_CREATIVE_DECRYPTOR.get()
+                Blocks.IRREALIUM_MACHINERY_CREATIVE_DECRYPTOR.get(),
+                Blocks.IRREALIUM_MACHINERY_CREATIVE_POWER_GENERATOR.get()
         );
         this.tag(BlockTags.IRREALIUM_STRUCTURE_CORNER).add(
                 Blocks.REALITY_BREAKTHROUGH_REACTOR_CONTROLLER.get(),