+++ src/main/java/site/siredvin/progressiveperipherals/data/ItemTagsProvider.java
@@ -31,7 +31,8 @@ public class ItemTagsProvider extends TagsProvider<Item> {
                 Blocks.IRREALIUM_MACHINERY_STORAGE.get().asItem(),
                 Blocks.REALITY_BREAKTHROUGH_REACTOR_CONTROLLER.get().asItem(),
                 Blocks.REALITY_BREAKTHROUGH_EXTRACTOR_CONTROLLER.get().asItem(),
-                Blocks.IRREALIUM_MACHINERY_CREATIVE_DECRYPTOR.get().asItem()
+                Blocks.IRREALIUM_MACHINERY_CREATIVE_DECRYPTOR.get().asItem(),
+                Blocks.IRREALIUM_MACHINERY_CREATIVE_POWER_GENERATOR.get().asItem()
         );
         this.tag(ItemTags.IRREALIUM_STRUCTURE_CORNER).add(
                 Blocks.REALITY_BREAKTHROUGH_REACTOR_CONTROLLER.get().asItem(),